luvbit/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx>=0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.7
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]==2.6.1
SQLAlchemy==2.0.31
psycopg2-binary

# 衛星軌道計算用ライブラリ
skyfield==1.48
numpy==1.26.4
geopy==2.4.1

# 追加の数学・計算ライブラリ
scipy==1.11.4
pytz==2023.3